tiem sync updates

+++ Core/Inc/neom9n.h
@@ -81,7 +81,7 @@ typedef struct {
 void gps_init(NEOM9N_t *packet, SPI_HandleTypeDef *hspi);
 bool gps_has_fix(const NEOM9N_t *packet);
 void pack_gps_data(const NEOM9N_t *packet, uint8_t *buffer);
-void pack_time_data(const NEOM9N_t *packet, uint8_t *buffer);
+void pack_time_data(const NEOM9N_t *packet, uint64_t *buffer);
 double time_to_seconds(const NEOM9N_t *packet);
 NEOM9N_status_t read_nmea_gga(NEOM9N_t *packet, uint32_t max_wait);
 NEOM9N_status_t read_nmea_gll(NEOM9N_t *packet, uint32_t max_wait);
